refactor: Agregando utils para extraer los useEffects de MapComponent
