+++ styles/pages/mapComponent.module.scss
@@ -1,16 +1,24 @@
+/* En el archivo mapComponent.module.scss */
+
 .container {
-    display: flex;
-    width: 100%;
-    height: 100vh;
-    justify-content: center;  
-    align-items: center;     
-  }
-  
-  #map {
-    width: 100%;
-    height: 100%;
-  }
-  .geocontainers{
+  display: flex;
+  width: 100%;
+  height: 100vh;
+  justify-content: center;
+  align-items: center;
+}
+
+.mapContainer {
+  position: relative;
+}
+
+#geocoder-container {
+  position: absolute;
+  top: 10px;
+  right: 10px;
+}
+
+.geocontainers {
   display: flex;
   flex-direction: column;
   width: 400px;
@@ -18,7 +26,7 @@
   gap: 20px;
 }
 
-  .ubicaciones{
-    background-color: aliceblue;
-    color: black;
-  }
+.ubicaciones {
+  background-color: aliceblue;
+  color: black;
+}
